emacs: add mechanism to prevent flickering of the echo area
When moving, Emacs:
- (t0) clears the echo area
- calls pre-command-hook
- (t1) updates the display
- calls post-command-hook
- calls our timer function to update the echo area
- (t2) updates the display

So even when our error message does not change (when moving left or
right on a line with an error), the buffer is cleared (t0),
drawn (t1), then updated and redrawn (t2), which creates flicker.

To avoid this, we undo the clearing in the pre-command-hook. For this,
we store the last shown error message in a variable.

(This cannot prevent flickering when moving between two consecutive
lines with each a different error, however this should be rarer.)