emacs: detect race condition with synchronous commands
merlin--waits-for-answer was not the right way to handle synchronous
commands.  I misunderstood emacs concurrency model *kof-kof*, calls are
not cooperative and share the same execution stack.
Two nested calls to merlin-send-command could result in priority
inversion and deadlock.

This situation is now detected and will make an assertion fail.  It was
happening when the lighter requested project information while a
synchronous command was being executed.  The lighter now checks that no
other command is being executed prior to updating its cache.

A better solution will be to make lighter queries asynchronous.