merlin-xref: make-point in the correct file (#1001)
When `loc.file` is not the current file, `(merlin/make-point pos)` tries to find the requested pos in the current file, may fail and return null. This null propagated to `merlin-xref--line` causes an error `Wrong type argument: integer-or-marker-p, nil`

So, let's switch to the correct file before invoking `(merlin/make-point)`.