Highlight xref matches in the target buffer more accurately
Use `xref-file-location` instead of `xref-buffer-location` since the
former gives line numbers and nicer display when there are multiple
matches on the same line.  Since Emacs 28, we have to partition the
line when populating the `summary` field.
Use the `match` face for highlighting if `xref-match` isn't
available (for Emacs 26 or older).