Detect operators and binding operators at point for xref
This will be used for `xref-find-references` (`M-?`), although that
command currently ignores the string and only uses the location.

If an identifier both succeeds and precedes point, prefer the former.
For instance, in `a+b` with the cursor on `+`, use `+` as the
idenfier; with the cursor on `b`, use `b`.